Adjust cluster covariance calculation to use new SACC file tracers (#92)
* Requiring sacc >= 0.9. This includes the new cluster changes and has no breaking changes.

* Removing old sacc files

* Updating the example notebook and checking in a new mock sacc file.

* Modified load_from_sacc to use the new data structure. Only significant change here is now richnesses are being saved into sacc in log base 10. I updated the code to read this in so the existing behavior doesnt change
CovarianceClusterCounts is now overriding _build_matrix_from_blocks and has been renamed. The reason is because before the tracers were unique per data point. This is not the case (we have the same tracers for cluster mass and cluster counts). We need to therefore specify the data_type we care about when building the final covariance matrix, otherwise we will populate other parts of the covariance matrix (e.g. cl mass) with the cl counts covariance.

* Added CovarianceClusterMass covariance calculator as a placeholder for the cluster weak lensing mass covariance matrix. Currently this is not implemented anywhere and needs to be established. For now it's simply the standard deviation of the masses in a given bin. By the time TJPCov is reading the sacc file this data is just a single point, so for now TJPCov just reads the previously calculated diagonal covariance and returns the same value. This will be changed in the future.